Fixed duplicating images and videos on update
